pyproject.toml
[project]
name = "outlook-bot"
version = "1.0.0"
description = "A local automation tool for macOS that bridges Microsoft Outlook with LLMs like Gemini and OpenAI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "certifi>=2025.11.12",
    "customtkinter>=5.2.2",
    "google-genai>=1.56.0",
    "httpx>=0.28.0",
    "openai>=2.14.0",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.1.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
]

[project.scripts]
outlook-bot = "outlook_bot.cli:main"
outlook-bot-gui = "outlook_bot.gui.app:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.10",
    "ty>=0.0.5",
    "pyinstaller>=6.0.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "A",     # flake8-builtins
    "C4",    # flake8-comprehensions
    "SIM",   # flake8-simplify
    "TCH",   # flake8-type-checking
    "RUF",   # ruff-specific rules
    "S",     # flake8-bandit (security)
    "PT",    # flake8-pytest-style
    "PIE",   # flake8-pie
    "T20",   # flake8-print (catch stray prints)
    "LOG",   # flake8-logging
    "G",     # flake8-logging-format
    "TID",   # flake8-tidy-imports
    "RSE",   # flake8-raise
    "RET",   # flake8-return
]
ignore = [
    "S603",   # subprocess-without-shell-check (we use osascript intentionally)
    "S607",   # start-process-with-partial-path (osascript is system binary)
    "T201",   # print (we use print for CLI output intentionally)
    "RET504", # unnecessary-assign (sometimes clearer with named returns)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S106", "S108"]

[tool.ruff.lint.isort]
known-first-party = ["outlook_bot"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ty.src]
include = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "-x",
]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["outlook_bot"]
omit = [
    "*/gui/*",
    "*/diagnostics/*",
]

[tool.coverage.report]
show_missing = true
fail_under = 60
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "@overload",
    "@abstractmethod",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
